* Entitlements for simulator - base Plist was extracted from InfoPlist class for generic Plist parsing and value placeholder substitution(with values from property file); - Simulator launch will use entitlements if provided in robovm.xml - changed way how team ID is generated for simulator launch. Instead of random string for each launch it will now: checks for `teamID` key in properties. If not -- will generate MD5 from bundle id -- to keep it same each launch. - removed hardcoded sim entitlement `keychain-access-groups` as entitlements can now be specified by user - for entitlements for device target properties are used as well. this will allow to use placeholders if required - removed bundled Entitlement.plist as it useless (empty)
